+++ nextjs/components/GamaStatusBanner/OteDisplay.tsx
@@ -1,0 +1,20 @@
+function OteDisplay() {
+    return <span
+                style={{
+                    padding: "6px 16px",
+                    background: "rgba(0, 0, 0, 0.3)",
+                    border: "2px solid #000000",
+                    borderRadius: "20px",
+                    color: "#000000",
+                    fontSize: "0.9rem",
+                    fontWeight: 900,
+                    animation: "pulse 1.5s infinite",
+                    boxShadow: "0 0 10px rgba(0, 0, 0, 0.4)",
+                    textShadow: "0 0 5px rgba(255, 255, 255, 0.2)"
+                }}
+            >
+                王手
+            </span>
+}
+
+export default OteDisplay;
