+++ nextjs/components/Overlay/DefeatAnimation.tsx
